ci: fix reload intent test linkage
